pyproject.toml
[project]
name = "vcut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.7.1",
    "jupyterlab>=4.4.2",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "scikit-image>=0.25.2",
]
